Create connection types
