+++ src/cow-react/modules/wallet/web3-react/hooks/useWalletInfo.ts
@@ -4,7 +4,7 @@ import useENSName from '@src/hooks/useENSName'
 import { UNSUPPORTED_WC_WALLETS } from 'constants/index'
 import { getConnection } from '@cow/modules/wallet/web3-react/utils/connection'
 import { SafeInfoResponse } from '@gnosis.pm/safe-service-client'
-import { ConnectionType } from '@cow/modules/wallet/web3-react/utils/connection/connections'
+import { ConnectionType } from '@cow/modules/wallet'
 import { useIsSmartContractWallet } from '@cow/common/hooks/useIsSmartContractWallet'
 import { useAtomValue } from 'jotai/utils'
 import { gnosisSafeAtom } from 'state/gnosisSafe/atoms'
