+++ src/cow-react/modules/wallet/web3-react/utils/connection/index.ts
@@ -1,8 +1,8 @@
 import { Connector } from '@web3-react/types'
+import { ConnectionType } from '@cow/modules/wallet'
 import {
   coinbaseWalletConnection,
-  Connection,
-  ConnectionType,
+  Connection,  
   fortmaticConnection,
   gnosisSafeConnection,
   injectedConnection,
