--- src/custom/state/connection/constants.ts
@@ -1,3 +1,0 @@
-import { ConnectionType } from '@cow/modules/wallet/web3-react/utils/connection/connections'
-
-export const BACKFILLABLE_WALLETS = [ConnectionType.INJECTED]
