+++ src/hooks/useOrderedConnections.ts
@@ -1,4 +1,4 @@
-import { ConnectionType } from '@cow/modules/wallet/web3-react/utils/connection/connections'
+import { ConnectionType } from '@cow/modules/wallet'
 import { getConnection } from '@cow/modules/wallet/web3-react/utils/connection'
 import { useMemo } from 'react'
 import { BACKFILLABLE_WALLETS } from 'state/connection/constants'
