+++ src/state/connection/constants.ts
@@ -1,4 +1,4 @@
-import { ConnectionType } from '@cow/modules/wallet/web3-react/utils/connection/connections'
+import { ConnectionType } from '@cow/modules/wallet'
 
 export const BACKFILLABLE_WALLETS = [
   ConnectionType.INJECTED,
