+++ src/state/connection/reducer.ts
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit'
-import { ConnectionType } from '@cow/modules/wallet/web3-react/utils/connection/connections'
+import { ConnectionType } from '@cow/modules/wallet'
 
 export interface ConnectionState {
   errorByConnectionType: Record<ConnectionType, string | undefined>
